Moved functions Evaluate_VB and Evaluate_dVB and other functions related to evaluation of potential and forces for indistinguishable particles to ipi.utils.exchange
